# api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PudimFinance API",
    "description": "Personal finance application API. Immutable double-entry ledger, event sourcing, real-time reporting, receipt scanning and price tracking.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "paths": {
    "/api/accounts": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Lists accounts with computed balances.",
        "operationId": "list_accounts",
        "responses": {
          "200": {
            "description": "List of accounts with balances",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountWithBalance"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Creates an account.",
        "operationId": "create_account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Account created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          },
          "400": {
            "description": "Invalid account payload"
          }
        }
      }
    },
    "/api/accounts/{id}": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Fetches an account with its computed balance.",
        "operationId": "get_account",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Account UUID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountWithBalance"
                }
              }
            }
          },
          "404": {
            "description": "Account not found"
          }
        }
      },
      "put": {
        "tags": [
          "Accounts"
        ],
        "summary": "Updates an account.",
        "operationId": "update_account",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Account UUID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Account updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          },
          "400": {
            "description": "Invalid account payload"
          },
          "404": {
            "description": "Account not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Accounts"
        ],
        "summary": "Deletes an account.",
        "operationId": "delete_account",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Account UUID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Account deleted"
          },
          "404": {
            "description": "Account not found"
          },
          "409": {
            "description": "Account is in use"
          }
        }
      }
    },
    "/api/accounts/{id}/adjust": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Adjusts an account to a target balance through a balanced equity posting.",
        "operationId": "adjust_account",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Account UUID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountAdjustmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Account adjustment recorded"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Account": {
        "type": "object",
        "description": "A chart-of-accounts account.",
        "required": [
          "id",
          "name",
          "type",
          "account_kind",
          "created_at"
        ],
        "properties": {
          "account_kind": {
            "type": "string",
            "description": "User-facing kind such as `bank`, `cash`, `card`, `loan`, `investment`, or a\nsystem kind (`income`/`expense`/`equity`/`other`)."
          },
          "closing_day": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Closing day of the monthly billing cycle (credit cards only, 1-31)."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Created at."
          },
          "credit_limit": {
            "type": [
              "string",
              "null"
            ],
            "description": "Credit limit (credit cards only)."
          },
          "due_day": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Payment due day of the monthly billing cycle (credit cards only, 1-31)."
          },
          "icon": {
            "type": [
              "string",
              "null"
            ],
            "description": "Stable icon identifier selected for this account."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Account ID."
          },
          "name": {
            "type": "string",
            "description": "Account display name (e.g., \"Cash\")."
          },
          "parent_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Optional parent account."
          },
          "type": {
            "type": "string",
            "description": "`asset`, `liability`, `equity`, `income`, or `expense`."
          }
        }
      },
...
      "AccountWithBalance": {
        "type": "object",
        "description": "Account joined with its current computed balance from ledger entries.",
        "required": [
          "id",
          "name",
          "type",
          "account_kind",
          "created_at",
          "balance",
          "transaction_count"
        ],
        "properties": {
          "account_kind": {
            "type": "string",
            "description": "User-facing kind such as `bank`, `cash`, `card`, `loan`, `investment`, or a\nsystem kind (`income`/`expense`/`equity`/`other`)."
          },
          "balance": {
            "type": "string",
            "description": "Balance = SUM(debit_amount) − SUM(credit_amount) across ledger entries.\nPositive for asset/expense accounts, negative for liability/income/equity\nunder standard double-entry semantics."
          },
          "closing_day": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Closing day of the monthly billing cycle (credit cards only, 1-31)."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Created at."
          },
          "credit_limit": {
            "type": [
              "string",
              "null"
            ],
            "description": "Credit limit (credit cards only)."
          },
          "due_day": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Payment due day of the monthly billing cycle (credit cards only, 1-31)."
          },
          "icon": {
            "type": [
              "string",
              "null"
            ],
            "description": "Stable icon identifier selected for this account."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Account ID."
          },
          "name": {
            "type": "string",
            "description": "Account display name (e.g., \"Credit Card\")."
          },
          "parent_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Optional parent account."
          },
          "transaction_count": {
            "type": "integer",
            "format": "int64",
            "description": "Total number of ledger entries posting to this account."
          },
          "type": {
            "type": "string",
            "description": "`asset`, `liability`, `equity`, `income`, or `expense`."
          }
        }
      },
...
      "CreateAccountRequest": {
        "type": "object",
        "description": "Payload for creating a new account.",
        "required": [
          "name",
          "type"
        ],
        "properties": {
          "account_kind": {
            "type": [
              "string",
              "null"
            ],
            "description": "User-facing kind. When set, the accounting `type` is derived from it\n(`bank`, `cash`, and `investment` map to asset, `card` and `loan` to liability).",
            "example": "card"
          },
          "closing_day": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Closing day of the monthly billing cycle (credit cards only, 1-31)."
          },
          "credit_limit": {
            "type": [
              "string",
              "null"
            ],
            "description": "Credit limit (credit cards only)."
          },
          "due_day": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Payment due day of the monthly billing cycle (credit cards only, 1-31)."
          },
          "icon": {
            "type": [
              "string",
              "null"
            ],
            "description": "Stable icon identifier selected for this account."
          },
          "initial_balance": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional opening balance. It is posted against the balance-adjustment equity account."
          },
          "name": {
            "type": "string",
            "description": "Account display name (e.g., \"Nubank Credit Card\")."
          },
          "parent_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Optional parent account."
          },
          "type": {
            "type": "string",
            "description": "`asset`, `liability`, `equity`, `income`, or `expense`. Ignored when\n`account_kind` is provided (the type is derived from the kind).",
            "example": "liability"
          }
        }
      },
...
      "UpdateAccountRequest": {
        "type": "object",
        "description": "Payload for updating an existing account.",
        "required": [
          "name",
          "type"
        ],
        "properties": {
          "account_kind": {
            "type": [
              "string",
              "null"
            ],
            "description": "User-facing kind. When set, the accounting `type` is derived from it.",
            "example": "card"
          },
          "closing_day": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Closing day of the monthly billing cycle (credit cards only, 1-31)."
          },
          "credit_limit": {
            "type": [
              "string",
              "null"
            ],
            "description": "Credit limit (credit cards only)."
          },
          "due_day": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Payment due day of the monthly billing cycle (credit cards only, 1-31)."
          },
          "icon": {
            "type": [
              "string",
              "null"
            ],
            "description": "Stable icon identifier selected for this account."
          },
          "name": {
            "type": "string",
            "description": "Account display name."
          },
          "parent_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Optional parent account."
          },
          "type": {
            "type": "string",
            "description": "`asset`, `liability`, `equity`, `income`, or `expense`. Ignored when\n`account_kind` is provided (the type is derived from the kind).",
            "example": "liability"
          }
        }
      },
...
    }
  },
  "security": [
    {}
  ],
...
}